--- src/modules/application/application.schema.ts
@@ -1,2 +1,0 @@
-export * from "./schema/application.request.schema";
-export * from "./schema/application.response.schema";
