+++ src/modules/myspace/applications.query.ts
@@ -1,0 +1,8 @@
+import { findVolunteerApplicationsByApplicantId } from "../volunteer/post-volunteer/post-volunteer.query";
+
+export type MySpaceVolunteerApplication =
+  Awaited<ReturnType<typeof findVolunteerApplicationsByApplicantId>>[number];
+
+export async function findMyVolunteerApplications(userId: string) {
+  return findVolunteerApplicationsByApplicantId(userId);
+}
