+++ src/modules/volunteer/post-volunteer/post-volunteer.query.ts
@@ -50,8 +50,10 @@ import {
 
 const ACTIVE_VOLUNTEER_APPLICATION_STATUSES = [
   "SUBMITTED",
-  "APPROVAL",
-  "ACCEPTED",
+  "UNDER_REVIEW",
+  "APPROVED",
+  "CONFIRMED",
+  "COMPLETED",
 ] as const;
 
 type VolunteerCategoryRow = typeof volunteerCategory.$inferSelect;
@@ -61,6 +63,7 @@ type VolunteerLocationRow = {
   name: string;
 };
 type VolunteerOpportunityRow = typeof volunteerOpportunity.$inferSelect;
+type VolunteerApplicationRow = typeof volunteerApplication.$inferSelect;
 type VolunteerOpportunitySaveInsert =
   typeof volunteerOpportunitySave.$inferInsert;
 type VolunteerRoleRow = typeof volunteerRole.$inferSelect;
@@ -165,10 +168,51 @@ export type VolunteerOpportunityDetail = {
     viewerApplied: boolean;
   }>;
 };
+export type VolunteerApplicationDetail = {
+  id: string;
+  opportunity: {
+    id: string;
+    title: string;
+    coverImageKey: string;
+    applicationDeadline: string;
+    status: VolunteerOpportunityRow["status"];
+    category: VolunteerReference;
+    location: VolunteerReference;
+  };
+  role: {
+    id: string;
+    title: string;
+  };
+  availability: string;
+  relevantExperience: string;
+  supportingDocumentKeys: string[];
+  status: VolunteerApplicationRow["status"];
+  createdAt: string;
+  updatedAt: string;
+};
 type VolunteerOpportunitiesListResult = {
   opportunities: VolunteerOpportunityListItem[];
   pagination: CursorPagination;
 };
+type VolunteerApplicationTarget = {
+  opportunityId: string;
+  opportunityTitle: string;
+  coverImageKey: string;
+  categoryId: string;
+  categoryName: string;
+  cityId: string;
+  cityName: string;
+  roleId: string;
+  roleTitle: string;
+  createdBy: string;
+  applicationDeadline: string;
+  status: VolunteerOpportunityRow["status"];
+  publishedAt: string | null;
+};
+type VolunteerApplicationOpportunityTarget = Omit<
+  VolunteerApplicationTarget,
+  "roleId" | "roleTitle"
+>;
 
 const CAMBODIA_NORMALIZED_NAME = env.VOLUNTEER_COUNTRY_NORMALIZED_NAME;
 const volunteerRoleSearch = aliasedTable(
@@ -323,12 +367,91 @@ export async function findVolunteerLocationById(
   return locationRow ?? null;
 }
 
+export async function findVolunteerOpportunityApplicationTargetById(
+  opportunityId: string,
+): Promise<VolunteerApplicationOpportunityTarget | null> {
+  const [row] = await db
+    .select({
+      opportunityId: volunteerOpportunity.id,
+      opportunityTitle: volunteerOpportunity.title,
+      coverImageKey: volunteerOpportunity.coverImageKey,
+      categoryId: volunteerCategory.id,
+      categoryName: volunteerCategory.name,
+      cityId: city.id,
+      cityName: city.name,
+      createdBy: volunteerOpportunity.createdBy,
+      applicationDeadline: volunteerOpportunity.applicationDeadline,
+      status: volunteerOpportunity.status,
+      publishedAt: volunteerOpportunity.publishedAt,
+    })
+    .from(volunteerOpportunity)
+    .innerJoin(
+      volunteerCategory,
+      eq(volunteerCategory.id, volunteerOpportunity.categoryId),
+    )
+    .innerJoin(city, eq(city.id, volunteerOpportunity.cityId))
+    .where(eq(volunteerOpportunity.id, opportunityId))
+    .limit(1);
+
+  return row ?? null;
+}
+
+export async function findVolunteerApplicationTargetByRoleId(
+  roleId: string,
+): Promise<VolunteerApplicationTarget | null> {
+  const [row] = await db
+    .select({
+      opportunityId: volunteerOpportunity.id,
+      opportunityTitle: volunteerOpportunity.title,
+      coverImageKey: volunteerOpportunity.coverImageKey,
+      categoryId: volunteerCategory.id,
+      categoryName: volunteerCategory.name,
+      cityId: city.id,
+      cityName: city.name,
+      roleId: volunteerRole.id,
+      roleTitle: volunteerRole.title,
+      createdBy: volunteerOpportunity.createdBy,
+      applicationDeadline: volunteerOpportunity.applicationDeadline,
+      status: volunteerOpportunity.status,
+      publishedAt: volunteerOpportunity.publishedAt,
+    })
+    .from(volunteerRole)
+    .innerJoin(
+      volunteerOpportunity,
+      eq(volunteerOpportunity.id, volunteerRole.opportunityId),
+    )
+    .innerJoin(
+      volunteerCategory,
+      eq(volunteerCategory.id, volunteerOpportunity.categoryId),
+    )
+    .innerJoin(city, eq(city.id, volunteerOpportunity.cityId))
+    .where(eq(volunteerRole.id, roleId))
+    .limit(1);
+
+  return row ?? null;
+}
+
 export type CreateVolunteerOpportunityInput =
   CreateVolunteerOpportunityBodyInput & {
     createdBy: string;
     category: VolunteerReference;
     location: VolunteerReference;
   };
+
+type CreateVolunteerApplicationInput = {
+  roleId: string;
+  applicantId: string;
+  opportunityId: string;
+  opportunityTitle: string;
+  coverImageKey: string;
+  applicationDeadline: string;
+  category: VolunteerReference;
+  location: VolunteerReference;
+  roleTitle: string;
+  availability: string;
+  relevantExperience: string;
+  supportingDocumentKeys: string[];
+};
 
 type VolunteerOpportunityListRow = VolunteerOpportunityBaseRow & {
   applicationCount: number;
@@ -353,6 +476,35 @@ function hydrateVolunteerOpportunityListItem(
     viewerSave: row.viewerSave,
     category: row.category,
     location: row.location,
+  };
+}
+
+function hydrateVolunteerApplication(
+  application: VolunteerApplicationRow,
+  roleTitle: string,
+  opportunity: {
+    id: string;
+    title: string;
+    coverImageKey: string;
+    applicationDeadline: string;
+    status: VolunteerOpportunityRow["status"];
+    category: VolunteerReference;
+    location: VolunteerReference;
+  },
+): VolunteerApplicationDetail {
+  return {
+    id: application.id,
+    opportunity,
+    role: {
+      id: application.roleId,
+      title: roleTitle,
+    },
+    availability: application.availability,
+    relevantExperience: application.relevantExperience,
+    supportingDocumentKeys: application.supportingDocumentKeys as string[],
+    status: application.status,
+    createdAt: application.createdAt,
+    updatedAt: application.updatedAt,
   };
 }
 
@@ -665,7 +817,7 @@ async function getAcceptedApplicationCountsByOpportunityIds(
     .where(
       and(
         inArray(volunteerApplication.opportunityId, uniqueOpportunityIds),
-        eq(volunteerApplication.status, "ACCEPTED"),
+        eq(volunteerApplication.status, "CONFIRMED"),
       ),
     )
     .groupBy(volunteerApplication.opportunityId);
@@ -1261,6 +1413,85 @@ export async function createVolunteerOpportunity(
       new Set(),
     );
   });
+}
+
+export async function createVolunteerApplication(
+  data: CreateVolunteerApplicationInput,
+): Promise<VolunteerApplicationDetail> {
+  return db.transaction(async (tx) => {
+    const [application] = await tx
+      .insert(volunteerApplication)
+      .values({
+        opportunityId: data.opportunityId,
+        roleId: data.roleId,
+        applicantId: data.applicantId,
+        availability: data.availability,
+        relevantExperience: data.relevantExperience,
+        supportingDocumentKeys: data.supportingDocumentKeys,
+        status: "SUBMITTED",
+      })
+      .returning();
+
+    return hydrateVolunteerApplication(application, data.roleTitle, {
+      id: data.opportunityId,
+      title: data.opportunityTitle,
+      coverImageKey: data.coverImageKey,
+      applicationDeadline: data.applicationDeadline,
+      status: "PUBLISHED",
+      category: data.category,
+      location: data.location,
+    });
+  });
+}
+
+export async function findVolunteerApplicationsByApplicantId(
+  applicantId: string,
+): Promise<VolunteerApplicationDetail[]> {
+  const rows = await db
+    .select({
+      application: volunteerApplication,
+      roleTitle: volunteerRole.title,
+      opportunityId: volunteerOpportunity.id,
+      opportunityTitle: volunteerOpportunity.title,
+      coverImageKey: volunteerOpportunity.coverImageKey,
+      applicationDeadline: volunteerOpportunity.applicationDeadline,
+      opportunityStatus: volunteerOpportunity.status,
+      categoryId: volunteerCategory.id,
+      categoryName: volunteerCategory.name,
+      cityId: city.id,
+      cityName: city.name,
+    })
+    .from(volunteerApplication)
+    .innerJoin(volunteerRole, eq(volunteerRole.id, volunteerApplication.roleId))
+    .innerJoin(
+      volunteerOpportunity,
+      eq(volunteerOpportunity.id, volunteerApplication.opportunityId),
+    )
+    .innerJoin(
+      volunteerCategory,
+      eq(volunteerCategory.id, volunteerOpportunity.categoryId),
+    )
+    .innerJoin(city, eq(city.id, volunteerOpportunity.cityId))
+    .where(eq(volunteerApplication.applicantId, applicantId))
+    .orderBy(desc(volunteerApplication.createdAt));
+
+  return rows.map((row) =>
+    hydrateVolunteerApplication(row.application, row.roleTitle, {
+      id: row.opportunityId,
+      title: row.opportunityTitle,
+      coverImageKey: row.coverImageKey,
+      applicationDeadline: row.applicationDeadline,
+      status: row.opportunityStatus,
+      category: {
+        id: row.categoryId,
+        name: row.categoryName,
+      },
+      location: {
+        id: row.cityId,
+        name: row.cityName,
+      },
+    }),
+  );
 }
 
 export async function saveVolunteerOpportunityForUser(
