+++ src/modules/volunteer/post-volunteer/post-volunteer.router.ts
@@ -6,6 +6,8 @@ import {
 } from "../../../middlewares/auth.middleware";
 import { authProtectedErrorResponseSchema } from "../../auth/auth.schema";
 import {
+  createVolunteerApplicationResponseSchema,
+  createVolunteerApplicationSchema,
   createVolunteerCategoryResponseSchema,
   createVolunteerCategorySchema,
   createVolunteerOpportunitySchema,
@@ -17,6 +19,8 @@ import {
   getVolunteerOpportunitiesResponseSchema,
   getVolunteerOpportunitiesQuerySchema,
   getSavedVolunteerOpportunitiesQuerySchema,
+  presignVolunteerApplicationDocumentUploadResponseSchema,
+  presignVolunteerApplicationDocumentUploadSchema,
   presignVolunteerOpportunityCoverUploadResponseSchema,
   presignVolunteerOpportunityCoverUploadSchema,
   saveVolunteerOpportunityResponseSchema,
@@ -25,6 +29,7 @@ import {
   volunteerValidationErrorResponseSchema,
 } from "./post-volunteer.schema";
 import {
+  handleCreateVolunteerApplication,
   handleCreateVolunteerCategory,
   handleCreateVolunteerOpportunity,
   handleGetSavedVolunteerOpportunities,
@@ -32,6 +37,7 @@ import {
   handleGetVolunteerLocations,
   handleGetVolunteerOpportunity,
   handleGetVolunteerOpportunities,
+  handlePresignVolunteerApplicationDocumentUpload,
   handlePresignVolunteerOpportunityCoverUpload,
   handleSaveVolunteerOpportunity,
   handleUnsaveVolunteerOpportunity,
@@ -235,6 +241,81 @@ const presignVolunteerOpportunityCoverUploadRoute = createRoute({
       content: {
         "application/json": {
           schema: authProtectedErrorResponseSchema,
+        },
+      },
+    },
+    500: {
+      description: "Internal server error",
+      content: {
+        "application/json": {
+          schema: volunteerOperationErrorResponseSchema,
+        },
+      },
+    },
+  },
+});
+
+const presignVolunteerApplicationDocumentUploadRoute = createRoute({
+  method: "post",
+  path: "/applications/document/presign",
+  tags: ["Volunteer Post"],
+  middleware: [requireAccessToken],
+  security: [{ BearerAuth: [] }],
+  request: {
+    body: {
+      content: {
+        "application/json": {
+          schema: presignVolunteerApplicationDocumentUploadSchema,
+        },
+      },
+    },
+  },
+  responses: {
+    200: {
+      description: "Presigned volunteer application document upload URL generated",
+      content: {
+        "application/json": {
+          schema: presignVolunteerApplicationDocumentUploadResponseSchema,
+        },
+      },
+    },
+    400: {
+      description: "Validation failed",
+      content: {
+        "application/json": {
+          schema: volunteerValidationErrorResponseSchema,
+        },
+      },
+    },
+    401: {
+      description: "Unauthorized",
+      content: {
+        "application/json": {
+          schema: authProtectedErrorResponseSchema,
+        },
+      },
+    },
+    403: {
+      description: "Forbidden",
+      content: {
+        "application/json": {
+          schema: authProtectedErrorResponseSchema,
+        },
+      },
+    },
+    404: {
+      description: "Volunteer opportunity not found",
+      content: {
+        "application/json": {
+          schema: volunteerOperationErrorResponseSchema,
+        },
+      },
+    },
+    409: {
+      description: "Volunteer application already exists",
+      content: {
+        "application/json": {
+          schema: volunteerOperationErrorResponseSchema,
         },
       },
     },
@@ -614,6 +695,81 @@ const createVolunteerOpportunityRoute = createRoute({
   },
 });
 
+const createVolunteerApplicationRoute = createRoute({
+  method: "post",
+  path: "/applications",
+  tags: ["Volunteer Post"],
+  middleware: [requireAccessToken],
+  security: [{ BearerAuth: [] }],
+  request: {
+    body: {
+      content: {
+        "application/json": {
+          schema: createVolunteerApplicationSchema,
+        },
+      },
+    },
+  },
+  responses: {
+    201: {
+      description: "Volunteer application submitted",
+      content: {
+        "application/json": {
+          schema: createVolunteerApplicationResponseSchema,
+        },
+      },
+    },
+    400: {
+      description: "Validation failed",
+      content: {
+        "application/json": {
+          schema: volunteerValidationErrorResponseSchema,
+        },
+      },
+    },
+    401: {
+      description: "Unauthorized",
+      content: {
+        "application/json": {
+          schema: authProtectedErrorResponseSchema,
+        },
+      },
+    },
+    403: {
+      description: "Forbidden",
+      content: {
+        "application/json": {
+          schema: authProtectedErrorResponseSchema,
+        },
+      },
+    },
+    404: {
+      description: "Volunteer role not found",
+      content: {
+        "application/json": {
+          schema: volunteerOperationErrorResponseSchema,
+        },
+      },
+    },
+    409: {
+      description: "Duplicate volunteer application",
+      content: {
+        "application/json": {
+          schema: volunteerOperationErrorResponseSchema,
+        },
+      },
+    },
+    500: {
+      description: "Internal server error",
+      content: {
+        "application/json": {
+          schema: volunteerOperationErrorResponseSchema,
+        },
+      },
+    },
+  },
+});
+
 postVolunteerRouter.openapi(getVolunteerCategoriesRoute, async (c) => {
   return handleGetVolunteerCategories(c) as any;
 });
@@ -632,6 +788,14 @@ postVolunteerRouter.openapi(
   async (c) => {
     const data = c.req.valid("json");
     return handlePresignVolunteerOpportunityCoverUpload(c, data) as any;
+  },
+);
+
+postVolunteerRouter.openapi(
+  presignVolunteerApplicationDocumentUploadRoute,
+  async (c) => {
+    const data = c.req.valid("json");
+    return handlePresignVolunteerApplicationDocumentUpload(c, data) as any;
   },
 );
 
@@ -663,4 +827,9 @@ postVolunteerRouter.openapi(unsaveVolunteerOpportunityRoute, async (c) => {
 postVolunteerRouter.openapi(createVolunteerOpportunityRoute, async (c) => {
   const data = c.req.valid("json");
   return handleCreateVolunteerOpportunity(c, data) as any;
+});
+
+postVolunteerRouter.openapi(createVolunteerApplicationRoute, async (c) => {
+  const data = c.req.valid("json");
+  return handleCreateVolunteerApplication(c, data) as any;
 });
