+++ src/modules/volunteer/post-volunteer/post-volunteer.service.ts
@@ -1,12 +1,18 @@
 import type { Context } from "hono";
 import { POSTGRES_UNIQUE_VIOLATION } from "../../../db/constants";
-import { presignVolunteerCoverUpload } from "../../uploads/uploads.service";
+import {
+  presignVolunteerApplicationDocumentUpload,
+  presignVolunteerCoverUpload,
+} from "../../uploads/uploads.service";
 import { getAuthUserId } from "../../auth/utils/get-auth";
 import {
+  createVolunteerApplication,
   createVolunteerCategory,
   createVolunteerOpportunity,
   findActiveVolunteerCategoryById,
+  findVolunteerApplicationTargetByRoleId,
   findVolunteerLocationById,
+  findVolunteerOpportunityApplicationTargetById,
   getSavedVolunteerOpportunities,
   getVolunteerCategories,
   getVolunteerOpportunityById,
@@ -16,11 +22,13 @@ import {
   unsaveVolunteerOpportunityForUser,
 } from "./post-volunteer.query";
 import type {
+  CreateVolunteerApplicationBodyInput,
   CreateVolunteerCategoryBodyInput,
   CreateVolunteerOpportunityBodyInput,
   GetSavedVolunteerOpportunitiesQuery,
   GetVolunteerOpportunityParams,
   GetVolunteerOpportunitiesQuery,
+  PresignVolunteerApplicationDocumentUploadPayload,
   PresignVolunteerOpportunityCoverUploadPayload,
 } from "./post-volunteer.schema";
 
@@ -28,6 +36,8 @@ const VOLUNTEER_CATEGORY_SLUG_UNIQUE_INDEX =
   "volunteer_category_slug_unique_idx";
 const VOLUNTEER_CATEGORY_NAME_UNIQUE_INDEX =
   "volunteer_category_name_unique_idx";
+const VOLUNTEER_APPLICATION_APPLICANT_ROLE_UNIQUE_INDEX =
+  "volunteer_application_applicant_role_active_unique_idx";
 
 type PostgresErrorLike = {
   code?: string;
@@ -62,6 +72,38 @@ function normalizeOwnedCoverImageKey(
   return normalizeOwnedObjectKey(`volunteer-covers/${userId}/`, coverImageKey);
 }
 
+function normalizeOwnedSupportingDocumentKey(
+  opportunityId: string,
+  applicantId: string,
+  supportingDocumentKey: string,
+): string | null {
+  return normalizeOwnedObjectKey(
+    `volunteer-applicant/supporting-docs/${opportunityId}/${applicantId}/`,
+    supportingDocumentKey,
+  );
+}
+
+function normalizeOwnedSupportingDocumentKeys(
+  opportunityId: string,
+  applicantId: string,
+  supportingDocumentKeys: string[],
+): string[] | null {
+  const normalizedKeys = supportingDocumentKeys.map((key) =>
+    normalizeOwnedSupportingDocumentKey(opportunityId, applicantId, key),
+  );
+
+  if (normalizedKeys.some((key) => key === null)) {
+    return null;
+  }
+
+  const resolvedNormalizedKeys = normalizedKeys as string[];
+  if (new Set(resolvedNormalizedKeys).size !== resolvedNormalizedKeys.length) {
+    return null;
+  }
+
+  return resolvedNormalizedKeys;
+}
+
 function normalizeOwnedObjectKey(
   expectedPrefix: string,
   objectKey: string,
@@ -84,6 +126,39 @@ function normalizeOwnedObjectKey(
   }
 
   return normalizedKey;
+}
+
+function resolveVolunteerApplicationTargetError(
+  target:
+    | {
+        createdBy: string;
+        applicationDeadline: string;
+        status: string;
+        publishedAt: string | null;
+      }
+    | null,
+  applicantId: string,
+  notFoundError: string,
+) {
+  if (!target || target.status !== "PUBLISHED" || !target.publishedAt) {
+    return { status: 404 as const, error: notFoundError };
+  }
+
+  if (target.createdBy === applicantId) {
+    return {
+      status: 400 as const,
+      error: "You cannot apply to your own volunteer opportunity",
+    };
+  }
+
+  if (Date.parse(target.applicationDeadline) <= Date.now()) {
+    return {
+      status: 400 as const,
+      error: "Application deadline has passed or been reached",
+    };
+  }
+
+  return null;
 }
 
 export async function handleGetVolunteerCategories(c: Context) {
@@ -321,6 +396,48 @@ export async function handlePresignVolunteerOpportunityCoverUpload(
   }
 }
 
+export async function handlePresignVolunteerApplicationDocumentUpload(
+  c: Context,
+  payload: PresignVolunteerApplicationDocumentUploadPayload,
+) {
+  const authResult = getAuthUserId(c);
+  if (!authResult.ok) {
+    return authResult.response;
+  }
+
+  try {
+    const target = await findVolunteerOpportunityApplicationTargetById(
+      payload.opportunityId,
+    );
+    const targetError = resolveVolunteerApplicationTargetError(
+      target,
+      authResult.userId,
+      "Volunteer opportunity not found",
+    );
+
+    if (targetError) {
+      return c.json({ ok: false, error: targetError.error }, targetError.status);
+    }
+
+    const uploads = payload.files.map((file) =>
+      presignVolunteerApplicationDocumentUpload({
+        opportunityId: payload.opportunityId,
+        applicantId: authResult.userId,
+        contentType: file.contentType,
+        fileSize: file.fileSize,
+      }),
+    );
+
+    return c.json({ ok: true, uploads }, 200);
+  } catch (error) {
+    console.error(
+      "Failed to generate volunteer application document upload URL",
+      error,
+    );
+    return c.json({ ok: false, error: "Failed to generate upload URL" }, 500);
+  }
+}
+
 export async function handleCreateVolunteerOpportunity(
   c: Context,
   data: CreateVolunteerOpportunityBodyInput,
@@ -376,6 +493,95 @@ export async function handleCreateVolunteerOpportunity(
     return c.json({ ok: true, opportunity }, 201);
   } catch (error) {
     console.error("Failed to create volunteer opportunity", error);
+    return c.json({ ok: false, error: "Internal server error" }, 500);
+  }
+}
+
+export async function handleCreateVolunteerApplication(
+  c: Context,
+  data: CreateVolunteerApplicationBodyInput,
+) {
+  const authResult = getAuthUserId(c);
+  if (!authResult.ok) {
+    return authResult.response;
+  }
+
+  try {
+    const target = await findVolunteerApplicationTargetByRoleId(data.roleId);
+    const targetError = resolveVolunteerApplicationTargetError(
+      target,
+      authResult.userId,
+      "Volunteer role not found",
+    );
+
+    if (targetError) {
+      return c.json({ ok: false, error: targetError.error }, targetError.status);
+    }
+
+    if (!target) {
+      throw new Error("Volunteer application target missing after validation");
+    }
+
+    const normalizedSupportingDocumentKeys = normalizeOwnedSupportingDocumentKeys(
+      target.opportunityId,
+      authResult.userId,
+      data.supportingDocumentKeys,
+    );
+
+    if (!normalizedSupportingDocumentKeys) {
+      return c.json(
+        {
+          ok: false,
+          error: "Validation failed",
+          issues: [
+            {
+              path: "supportingDocumentKeys",
+              message:
+                "supportingDocumentKeys must belong to current user and opportunity",
+            },
+          ],
+        },
+        400,
+      );
+    }
+
+    const application = await createVolunteerApplication({
+      roleId: data.roleId,
+      applicantId: authResult.userId,
+      opportunityId: target.opportunityId,
+      opportunityTitle: target.opportunityTitle,
+      coverImageKey: target.coverImageKey,
+      applicationDeadline: target.applicationDeadline,
+      category: {
+        id: target.categoryId,
+        name: target.categoryName,
+      },
+      location: {
+        id: target.cityId,
+        name: target.cityName,
+      },
+      roleTitle: target.roleTitle,
+      availability: data.availability,
+      relevantExperience: data.relevantExperience,
+      supportingDocumentKeys: normalizedSupportingDocumentKeys,
+    });
+
+    return c.json({ ok: true, application }, 201);
+  } catch (err) {
+    const error = getPostgresError(err);
+
+    if (
+      error?.code === POSTGRES_UNIQUE_VIOLATION &&
+      getPostgresConstraint(error) ===
+        VOLUNTEER_APPLICATION_APPLICANT_ROLE_UNIQUE_INDEX
+    ) {
+      return c.json(
+        { ok: false, error: "You have already applied to this role" },
+        409,
+      );
+    }
+
+    console.error("Failed to create volunteer application", err);
     return c.json({ ok: false, error: "Internal server error" }, 500);
   }
 }
