+++ src/modules/volunteer/post-volunteer/schema/opportunities.request.schema.ts
@@ -6,10 +6,14 @@ const VOLUNTEER_COVER_IMAGE_ALLOWED_CONTENT_TYPES = [
   "image/png",
   "image/webp",
 ] as const;
+const VOLUNTEER_APPLICATION_DOCUMENT_CONTENT_TYPE = "application/pdf";
+const MIN_VOLUNTEER_APPLICATION_DOCUMENT_COUNT = 1;
+const MAX_VOLUNTEER_APPLICATION_DOCUMENT_COUNT = 3;
 const MAX_VOLUNTEER_OPPORTUNITIES_PAGE_SIZE = 50;
 const DEFAULT_VOLUNTEER_OPPORTUNITIES_PAGE_SIZE = 10;
 
 export const VOLUNTEER_COVER_IMAGE_MAX_BYTES = 5 * 1024 * 1024;
+export const VOLUNTEER_APPLICATION_DOCUMENT_MAX_BYTES = 10 * 1024 * 1024;
 
 function normalizeText(value: string) {
   return value.trim().replace(/\s+/g, " ");
@@ -305,6 +309,44 @@ export const presignVolunteerOpportunityCoverUploadSchema = z
   })
   .openapi("PresignVolunteerOpportunityCoverUploadRequest");
 
+export const presignVolunteerApplicationDocumentUploadSchema = z
+  .object({
+    opportunityId: z
+      .string()
+      .trim()
+      .regex(VOLUNTEER_UUID_RE, "opportunityId must be a valid UUID"),
+    files: z
+      .array(
+        z.object({
+          contentType: z
+            .string()
+            .trim()
+            .toLowerCase()
+            .refine(
+              (value) => value === VOLUNTEER_APPLICATION_DOCUMENT_CONTENT_TYPE,
+              "Supporting document must be a PDF",
+            ),
+          fileSize: z
+            .number()
+            .int("fileSize must be an integer")
+            .positive("fileSize must be positive")
+            .max(
+              VOLUNTEER_APPLICATION_DOCUMENT_MAX_BYTES,
+              `fileSize must be <= ${VOLUNTEER_APPLICATION_DOCUMENT_MAX_BYTES / (1024 * 1024)} MB`,
+            ),
+        }),
+      )
+      .min(
+        MIN_VOLUNTEER_APPLICATION_DOCUMENT_COUNT,
+        `files must contain at least ${MIN_VOLUNTEER_APPLICATION_DOCUMENT_COUNT} files`,
+      )
+      .max(
+        MAX_VOLUNTEER_APPLICATION_DOCUMENT_COUNT,
+        `files must contain at most ${MAX_VOLUNTEER_APPLICATION_DOCUMENT_COUNT} files`,
+      ),
+  })
+  .openapi("PresignVolunteerApplicationDocumentUploadRequest");
+
 const createVolunteerOpportunityBaseSchema = z
   .object({
     categoryId: z.string().uuid("categoryId must be a valid UUID"),
@@ -397,8 +439,73 @@ export const createVolunteerOpportunitySchema =
     })
     .openapi("CreateVolunteerOpportunityRequest");
 
+export const createVolunteerApplicationSchema = z
+  .object({
+    roleId: z.string().uuid("roleId must be a valid UUID"),
+    availability: z
+      .string()
+      .transform((value) => normalizeText(value))
+      .pipe(
+        z
+          .string()
+          .min(1, "availability is required and must be 1..500 characters")
+          .max(500, "availability is required and must be 1..500 characters"),
+      ),
+    relevantExperience: z
+      .string()
+      .transform((value) => normalizeText(value))
+      .pipe(
+        z
+          .string()
+          .min(
+            1,
+            "relevantExperience is required and must be 1..5000 characters",
+          )
+          .max(
+            5000,
+            "relevantExperience is required and must be 1..5000 characters",
+          ),
+      ),
+    supportingDocumentKeys: z
+      .array(
+        z
+          .string()
+          .trim()
+          .min(1, "supportingDocumentKeys[] is required")
+          .max(600, "supportingDocumentKeys[] must be <= 600 characters"),
+      )
+      .min(
+        MIN_VOLUNTEER_APPLICATION_DOCUMENT_COUNT,
+        `supportingDocumentKeys must contain at least ${MIN_VOLUNTEER_APPLICATION_DOCUMENT_COUNT} files`,
+      )
+      .max(
+        MAX_VOLUNTEER_APPLICATION_DOCUMENT_COUNT,
+        `supportingDocumentKeys must contain at most ${MAX_VOLUNTEER_APPLICATION_DOCUMENT_COUNT} files`,
+      )
+      .superRefine((value, ctx) => {
+        const seen = new Set<string>();
+        value.forEach((item, index) => {
+          if (seen.has(item)) {
+            ctx.addIssue({
+              code: z.ZodIssueCode.custom,
+              path: [index],
+              message: "supportingDocumentKeys[] must not contain duplicates",
+            });
+            return;
+          }
+
+          seen.add(item);
+        });
+      }),
+  })
+  .openapi("CreateVolunteerApplicationRequest");
+
 export type PresignVolunteerOpportunityCoverUploadPayload = z.infer<
   typeof presignVolunteerOpportunityCoverUploadSchema
+>;
+
+export type PresignVolunteerApplicationDocumentUploadPayload = z.infer<
+  typeof presignVolunteerApplicationDocumentUploadSchema
 >;
 
 export const getVolunteerOpportunitiesQuerySchema = z
@@ -504,4 +611,8 @@ export type GetVolunteerOpportunityParams = z.infer<
 
 export type CreateVolunteerOpportunityBodyInput = z.infer<
   typeof createVolunteerOpportunitySchema
+>;
+
+export type CreateVolunteerApplicationBodyInput = z.infer<
+  typeof createVolunteerApplicationSchema
 >;
