+++ src/modules/volunteer/post-volunteer/schema/opportunities.response.schema.ts
@@ -68,6 +68,24 @@ const volunteerOpportunityReferenceSchema = z
   })
   .openapi("VolunteerOpportunityReference");
 
+const volunteerApplicationRoleResponseSchema = z
+  .object({
+    id: z.string(),
+    title: z.string(),
+  })
+  .openapi("VolunteerApplicationRoleResponse");
+
+const volunteerApplicationOpportunitySchema = z
+  .object({
+    id: z.string(),
+    title: z.string(),
+    coverImageKey: z.string(),
+    status: z.enum(["DRAFT", "PUBLISHED", "ARCHIVED", "CLOSED"]),
+    category: volunteerOpportunityReferenceSchema,
+    location: volunteerOpportunityReferenceSchema,
+  })
+  .openapi("VolunteerApplicationOpportunity");
+
 const volunteerOpportunityOrganizerResponseSchema = z
   .object({
     id: z.string(),
@@ -136,6 +154,35 @@ export const createVolunteerOpportunityResponseSchema = z
     opportunity: volunteerOpportunityResponseSchema,
   })
   .openapi("CreateVolunteerOpportunityResponse");
+
+export const volunteerApplicationResponseSchema = z
+  .object({
+    id: z.string(),
+    opportunity: volunteerApplicationOpportunitySchema,
+    role: volunteerApplicationRoleResponseSchema,
+    availability: z.string(),
+    relevantExperience: z.string(),
+    supportingDocumentKeys: z.array(z.string()),
+    status: z.enum([
+      "SUBMITTED",
+      "UNDER_REVIEW",
+      "APPROVED",
+      "DECLINED",
+      "CONFIRMED",
+      "COMPLETED",
+      "WITHDRAWN",
+    ]),
+    createdAt: z.string(),
+    updatedAt: z.string(),
+  })
+  .openapi("VolunteerApplicationResponse");
+
+export const createVolunteerApplicationResponseSchema = z
+  .object({
+    ok: z.literal(true),
+    application: volunteerApplicationResponseSchema,
+  })
+  .openapi("CreateVolunteerApplicationResponse");
 
 export const getVolunteerOpportunityResponseSchema = z
   .object({
